# Conflicts:
#	bin/main.js
#	bin/root/tools/deploy.js
#	bin/root/tools/resx.js
#	bin/root/tools/setFormCustomizable.js
#	package.json

+++ src/generator/Entity.ts
@@ -31,16 +31,18 @@ export class Entity {
         shell.ls(`${__dirname}/Entity/*.*`).forEach(function (file) {
             const split = file.split('/');
             const filename = split[split.length - 1];
-            const newfilename = filename.replace(/Entity/g, entityname);
-            shell.cp('-r', file, `src/${entityname}`);
-            shell.cp('-r', `src/${entityname}/${filename}`, `src/${entityname}/${newfilename}`);
-            shell.rm('-rf', `src/${entityname}/${filename}`);
-            shell.sed('-i', new RegExp('EntityLogicalName', 'ig'), answers.entityLogicalName, `src/${entityname}/${newfilename}`);
-            shell.sed('-i', new RegExp('Entity', 'g'), entityname, `src/${entityname}/${newfilename}`);
-            shell.sed('-i', new RegExp('entity', 'g'), entityname.charAt(0).toLowerCase() + entityname.slice(1), `src/${entityname}/${newfilename}`);
-            shell.sed('-i', new RegExp('<%= publisher %>', 'ig'), publisher, `src/${entityname}/${newfilename}`);
-            shell.sed('-i', new RegExp('<%= namespace %>', 'ig'), namespace, `src/${entityname}/${newfilename}`);
-            shell.exec(`git add src/${entityname}/${newfilename}`);
+            if (filename !== 'Entity.enum.ts') {
+                const newfilename = filename.replace(/Entity/g, entityname);
+                shell.cp('-r', file, `src/${entityname}`);
+                shell.cp('-r', `src/${entityname}/${filename}`, `src/${entityname}/${newfilename}`);
+                shell.rm('-rf', `src/${entityname}/${filename}`);
+                shell.sed('-i', new RegExp('EntityLogicalName', 'ig'), answers.entityLogicalName, `src/${entityname}/${newfilename}`);
+                shell.sed('-i', new RegExp('Entity', 'g'), entityname, `src/${entityname}/${newfilename}`);
+                shell.sed('-i', new RegExp('entity', 'g'), entityname.charAt(0).toLowerCase() + entityname.slice(1), `src/${entityname}/${newfilename}`);
+                shell.sed('-i', new RegExp('<%= publisher %>', 'ig'), publisher, `src/${entityname}/${newfilename}`);
+                shell.sed('-i', new RegExp('<%= namespace %>', 'ig'), namespace, `src/${entityname}/${newfilename}`);
+                shell.exec(`git add src/${entityname}/${newfilename}`);
+            }
         });
         const webpackConfigFile = shell.ls('webpack.config.js')[0];
         // eslint-disable-next-line max-len
